+++ app/src/main/java/com/nawl/carmaintenanceapp/model/entities/Vehicle.kt
@@ -6,6 +6,6 @@ import androidx.room.PrimaryKey
 
 @Entity (tableName = "vehicle")
 data class Vehicle(
-    @PrimaryKey(autoGenerate = true) val vehicleId: Int = 0,
+    @PrimaryKey(autoGenerate = true) val vehicleId: Long = 0,
     @ColumnInfo(name = "name") val name: String
 )
